+++ benchmarks/exp_hash.py
@@ -69,15 +69,17 @@ def _human_size(size: int) -> str:
 
 
 def _get_hasher(hash_algorithm: str) -> hashing.StreamingHashEngine:
-    # TODO: Once Python 3.9 support is deprecated revert to using `match`
-    if hash_algorithm == "sha256":
-        return memory.SHA256()
-    if hash_algorithm == "blake2":
-        return memory.BLAKE2()
-    if hash_algorithm == "blake3":
-        return memory.BLAKE3()
-
-    raise ValueError(f"Cannot convert {hash_algorithm} to a hash engine")
+    match hash_algorithm:
+        case "sha256":
+            return memory.SHA256()
+        case "blake2":
+            return memory.BLAKE2()
+        case "blake3":
+            return memory.BLAKE3()
+        case _:
+            raise ValueError(
+                f"Cannot convert {hash_algorithm} to a hash engine"
+            )
 
 
 def _generate_data(size: int) -> bytes: